% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessResults.R
\name{connect_to_results}
\alias{connect_to_results}
\title{Connect to a trial design SQLite database}
\usage{
connect_to_results(sqlite_filename)
}
\arguments{
\item{sqlite_filename}{path to trial design SQLite database}
}
\value{
an RSQLite database connection object
}
\description{
Connect to a trial design SQLite database
}
