% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccessResults.R
\name{fetch_result}
\alias{fetch_result}
\title{Get information from a trial design SQLite database}
\usage{
fetch_result(db_conn, a0, a1, b0, b1)
}
\arguments{
\item{db_conn}{RSQLite database connection object (connection to the trial design database)}

\item{a0}{entry A0 of your current contingency table}

\item{a1}{entry A1 of your current contingency table}

\item{b0}{entry B0 of your current contingency table}

\item{b1}{entry B1 of your current contingency table}
}
\value{
a list of values, including the size of the next stage ("BlockSize") and number of patients to allocate to A ("AAllocation")
}
\description{
Given your current contingency table, retrieve the size and treatment allocation of the next trial stage.
}
