% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{trial_mdp}
\alias{trial_mdp}
\title{Use TrialMDP to compute an optimal trial design}
\usage{
trial_mdp(
  n_patients,
  failure_cost,
  block_cost,
  sqlite_fname,
  min_size = 4L,
  block_incr = 2L,
  prior_a0 = 1,
  prior_a1 = 1,
  prior_b0 = 1,
  prior_b1 = 1,
  transition_dist = "beta_binom",
  test_statistic = "scaled_cmh",
  act_l = 0.2,
  act_u = 0.8,
  act_n = 7L
)
}
\arguments{
\item{n_patients}{the number of patients in the trial}

\item{failure_cost}{parameter representing the cost of patient failures}

\item{block_cost}{parameter representing the cost of each additional trial stage}

\item{sqlite_fname}{output filepath for trial design SQLite database}

\item{min_size}{minimum size for a trial stage. Default=4}

\item{block_incr}{require trial stage sizes to be multiples of this number. Default=2}

\item{prior_a0}{smoothing/pseudocount hyperparameter for computing transition probabilities. Default=1.0}

\item{prior_a1}{smoothing/pseudocount hyperparameter for computing transition probabilities. Default=1.0}

\item{prior_b0}{smoothing/pseudocount hyperparameter for computing transition probabilities. Default=1.0}

\item{prior_b1}{smoothing/pseudocount hyperparameter for computing transition probabilities. Default=1.0}

\item{transition_dist}{name of transition probability distribution. Default="beta_binom". We do not recommend changing this.}

\item{test_statistic}{name of test statistic for which to optimize. Default="scaled_cmh". We do not recommend changing this.}

\item{act_l}{smallest allocation fraction to treatment A. i.e., Phi = {act_l, ..., act_u}. Default=0.2}

\item{act_u}{largest allocation fraction to treatment A. i.e., Phi = {act_l, ..., act_u}. Default=0.8}

\item{act_n}{number of possible allocation fractions, uniformly spaced. i.e., |Phi| = act_n. Default=7}
}
\value{
None. Trial design is written to disk.
}
\description{
Given the number of patients, failure cost, and stage cost, compute an optimal trial design and save it to a SQLite database.
}
